pyproject.toml
[tool.poetry]
name = "fastapi-example-project"
version = "0.1.0"
description = "Example FastAPI project with core components."
authors = ["datadave <605409+datadavev@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
pytest = "^7.4.2"
pydantic-settings = "^2.0.3"
jinja2 = "^3.1.2"
github-webhooks-framework = {git = "https://github.com/datadavev/github-webhooks.git", rev = "main"}
json-logging = "^1.3.0"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
